+++ src/test/java/test/GetPetById.java
@@ -17,7 +17,7 @@ public class GetPetById {
 
     @Steps
     private PetEndpoint petEndpoint;
-    private long petId;
+    private int petId;
 
     @Before
     public void createPet() {
